+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* TBD:
+  * Pay by scanning a paper wallet (using camera or NFC)
 * v1.3.0:
   * Segwit support (native and backwards compatible addresses)
   * Hide testnet payment methods by default